Refactor ping command
Refactors ping command to follow the new pattern of including responses
in the command files.
